+++ src/ngrest/render/RenderCrud.php
@@ -115,7 +115,8 @@ class RenderCrud extends Render implements ViewContextInterface, RenderCrudInter
             'config' => $this->config,
             'isInline' => $this->getIsInline(),
             'modelSelection' => $this->getModelSelection(),
-            'relationCall' => $this->getRelationCall(), // this is currently only used for the curd_relation view file, there for split the RenderCrud into two sepeare renderes.
+            'hasActiveSelections' => count($this->config->getActiveSelections()),
+            'relationCall' => $this->getRelationCall(), // this is currently only used for the crud relation view file, there for split the RenderCrud into two sepeare renderes.
             'currentMenu' => Yii::$app->adminmenu->getApiDetail($this->getConfig()->getApiEndpoint()),
             'downloadAttributes' => $this->generateDownloadAttributes(),
         ], $this);